+++ src/contexts/daemon/infrastructure.rs
@@ -5,3 +5,5 @@ pub mod paths;
 pub mod pid;
 pub(super) mod protocol;
 pub(super) mod repo_id;
+pub(super) mod request_handler;
+pub(super) mod server_config;
